Generate auth artifacts for guest users
This enables users in the _data/private/hub/guest_users.yml list to access the
Hub.

Authenication includes are generated, but do not link to any pages. On top of
being a nice touch, they're necessary to avoid breaking the Server Side
Include directive in _layouts/bare.html, since there's no means of checking
the existence of an included file in SSI itself. We could write a small CGI
process to check this and generate the authentication include on the fly, but
that defeats the purpose of keeping the Hub otherwise all-static.