$http_x_forwarded_email => $authenticated_user
This is in anticipation of using Vagrant + Ansible to launch a local
development server running Nginx. Rather than using the `set` directive to
change the value of `$http_x_forwarded_email` directly in the configuration
for the development instance, updating the basic layout to use
`$authenticated_user` makes the Server Side Include directive a bit more
flexible and understandable.

The real reason I ran up against this: using the `set` directive to change a
value of a `$http_...` variable causes a segfault in Nginx < 1.7.5:
  http://nginx.org/en/CHANGES