Move deployment commands into ./go script
The command strings in deploy/fabfile.py were getting a bit unwieldy; this is
an attempt to abstract them in a more coherent way. Also, any tweaks to the
deployment commands will not require any `fab [stop|start]` invocations.

Another approach might be to add separate scripts in deploy/ for each
individual instance, rather than sticking them in ./go. Open to doing it that
way as well.