Add tests for untitled posts.
I added two new tests that check for empty post titles and titles that are literally set to "Untitled". The failure error will show you an array of URLs for posts that don't have titles.

I also added the coverage folder to `.gitignore` since it doesn't look like it's in the GitHub repo.