Fix nginx config; make 403 page style standalone
The previous changes in #323 didn't actually do the trick, as the 403 is
coming from the oauth2_proxy, not directly from nginx. This change allows the
auth.18f.gov virtual server to intercept the 403 from oauth2_proxy during
the sign-in process, after the user has successfully logged into MyUSA but
using an account that isn't authorized to access the Hub.

The `auth.conf` config uses a Server-Side Include to serve a 403 page with a
custom link. This custom link allows the user to automatically log out of
MyUSA, and will also redirect to the intended URL once the user has logged
back in using the correct account. By changing the Hub's application URL in
MyUSA to https://auth.18f.gov/, this configuration works for all 18F
properties authenticated using the Hub's oauth2_proxy (including
https://tock.18f.gov/ and https://pages-staging.18f.gov/).

Since an unauthenticated user will not be able to load scrips and stylesheets,
there is a new http_error layout that cuts most of these things and inlines
the minimum styles necessary to maintain a consistent look (error.scss).

Also adds a helpful link to check the veracity of the Shaxper quote.