Link to MyUSA swap accounts directly from 403 page
In #324, the link at the bottom of the page will take the user to the
oauth2_proxy's sign-in page, which will then take the user to MyUSA to swap
accounts before redirecting back to the intended page. This nginx config tweak
enables the link to be built such that it leads directly to MyUSA while
preserving the redirect to the intended page.