Fix /skills and /interests pages
At some point, and I'm not sure if it was a Jekyll upgrade or the move to turn
_pages into a collection, the scripts specified in _layouts/skills.html were
no longer inherited by _pages/skills.html and _pages/interests.html, causing
both pages and pages generated dynamically from them to be blank. Listing
the scripts directly in the page objects resolves the issue.

Took the liberty to reformat _includes/scripts.html as well.