Adding weekly intake standups
Lots of folks have been suggesting folks attend the weekly intake standup when they're new - great way to get to know the way we work and the projects that are coming in. Also worth mentioned - I'm not sure what this Tuesday 1:00 ET meeting is - nobody has said anything to be about it. Might be worth clarifying or deleting?